Update application properties and controller settings
Changed application properties to switch from festisounds database to postgres database. Updated FestivalController to adjust for these changes.

This adjustment was made in consideration for using Docker. A docker-compose file was also added to make use of Postgres in development environment. The adjustment will ensure the alignment and compatibility with the containerization plan. Also, in the FestivalController, the BASE_PATH is amended and the festivalRepo is instantiated in the controller's constructor, adhering to a better dependency injection practice.

Co-authored-by: Cornelis Verkade <verkadecornelis@gmail.com>
Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>