Refactor code structure for better readability and functionality
This commit involves several changes aimed at improving the code structure and enhance readability. Deleted `FestivalDTO` in the root `DTOs` directory and it's recreated under `Festival` module for better organization. The `SpotifyDataService` and `SpotifyDataController` have been moved to the `Modules` section, under `SpotifyData` folder along with its service implementation `SpotifyDataServiceImpl`.

Similar refactoring was applied to `GlobalErrorHandler` and `WebMvcConfiguration` which are moved to the `Core` section. Refactored `FestivalController` and `Festival` entity folders to reflect their belonging in the `Festival` module. The `ApplicationTestConfiguration` is updated as per the refactoring changes made in the services.

Lastly, copious changes made to `.gitignore` to exclude unnecessary files from the repository.

These changes are made in accordance with best practices for project organization, to enhance code readability, ease future maintenance, and bring uniformity across various modules of the project.

Co-authored-by: Cornelis Verkade <verkadecornelis@gmail.com>
Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>