Add RequestTest and ResponseTest for Festival module
New files, `RequestTest` and `ResponseTest`, have been added under `Festival` module in the `Requests` and `Responses` directories respectively. The addition of these files is part of the effort to improve code organization, structure, and maintainability. It will provide a convenient way of handling requests and responses in the `Festival` module, making the code cleaner and easier to read.

Co-authored-by: Cornelis Verkade <verkadecornelis@gmail.com>
Signed-off-by: Maximiliano Barrientos <elbodysathba@gmail.com>